refactor: split server dashboard pages
